+++ src/main/java/sk/nociar/jpacloner/PropertyFilters.java
@@ -2,12 +2,15 @@ package sk.nociar.jpacloner;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import sk.nociar.jpacloner.AbstractJpaExplorer.JpaClassInfo;
 import sk.nociar.jpacloner.graphs.PropertyFilter;
+
+import javax.persistence.AccessType;
 
 /**
  * Factory of various {@link PropertyFilter}s. Example:<br/>
@@ -48,9 +51,16 @@ public class PropertyFilters {
 		public boolean test(Object entity, String property) {
 			JpaClassInfo info = AbstractJpaExplorer.getClassInfo(entity);
 			if (info != null) {
-				Field field = info.getField(property);
-				if (field != null && field.getAnnotation(clazz) != null) {
-					return false;
+				if(info.getAccessType() == AccessType.PROPERTY) {
+					Method getter = info.getGetter(property);
+					if (getter != null && getter.getAnnotation(clazz) != null) {
+						return false;
+					}
+				} else {
+					Field field = info.getField(property);
+					if (field != null && field.getAnnotation(clazz) != null) {
+						return false;
+					}
 				}
 			}
 			return true;